+++ draft-packages/checkbox-group/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-checkbox-group@1.4.21...@kaizen/draft-checkbox-group@1.4.22) (2020-11-09)
+
+**Note:** Version bump only for package @kaizen/draft-checkbox-group
+
+
+
+
+
+## [1.4.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-checkbox-group@1.4.20...@kaizen/draft-checkbox-group@1.4.21) (2020-11-09)
+
+**Note:** Version bump only for package @kaizen/draft-checkbox-group
+
+
+
+
+
 ## [1.4.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-checkbox-group@1.4.19...@kaizen/draft-checkbox-group@1.4.20) (2020-10-21)
 
 **Note:** Version bump only for package @kaizen/draft-checkbox-group