+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.6...@kaizen/draft-modal@3.2.7) (2020-11-09)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [3.2.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.5...@kaizen/draft-modal@3.2.6) (2020-11-09)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [3.2.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.4...@kaizen/draft-modal@3.2.5) (2020-10-29)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [3.2.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.3...@kaizen/draft-modal@3.2.4) (2020-10-28)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [3.2.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@3.2.2...@kaizen/draft-modal@3.2.3) (2020-10-28)
 
 **Note:** Version bump only for package @kaizen/draft-modal